requirements.txt
stable-baselines3
gymnasium
matplotlib
shapely
dubins
pyyaml